contain them in a class
