+++ README.md
@@ -1,6 +1,12 @@
 # Description
 PyMoon is a module that can calculate your moonphase, based on which it gives you life suggestions, compatibility advices, and personality traits.
 
+To use:
+
+`from moon_phases import moon_phases`
+
+`print()`
+
 # Set Up
 First install pipenv if it's not installed on your computer.
