+++ src/package/__init__.py
@@ -1,0 +1,1 @@
+from moon_phases import PyMoon
